cerbero: Implement selecting of build type via a command-line argument
This also reports the currently-configured build type if
~/.cerbero/cerbero.cbc is used to set the debug variant.

Ported from the meson-1.8 branch. This was somehow missed in this
branch.

Note that 8d6511695c4433c66214fda6e36af0121f75b4e4 referred to the
`set_buildtype` function that this commit adds. That commit seems to
have been accidentally cherry-picked without this one.