meson: Use the correct libstdc++ in cerbero shell
By default, when a Cerbero shell is run, the MSYS shell adds its own
paths in the beginning of PATH. This causes applications to use the MSYS
libstdc++ when C++ applications are run. This is obviously broken since
that is incompatible with the libstdc++ that the application was
compiled with (the MinGW toolchain that we install into ~/cerbero/mingw)

Now we install a profile.d script that rewrites PATH to fix this.