Update README for Meson port and MSVC usage
With the Meson port, we can now build many recipes with MSVC, and we
will default to doing that on Windows. People can switch to MinGW-only
by adding 'novisualstudio' to variants in ~/.cerbero/cerbero.cbc

We always needed an MSVC compiler for building the directshow plugins,
and we need lib.exe from Visual Studio to generate import libraries
from DLLs for all recipes built with Autotools.

This change documents that correctly and bumps the requirement to
Visual Studio 2015, which is what we support in GStreamer. VS 2013 may
work, and support for it can be added if enough people ask for it.

Also fix the documentation about how to install Python 3. The default
installation path is in `C:\Program Files (x86)\Python36-32` which
breaks Autotools and other shell-based builds. Note that Meson does
not require this.